Rename the enums file to make it more meaningful.
